feat: create component for driveItems
